requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
sqlalchemy>=2.0.0
pydantic>=1.10.7
pandas>=1.5.3
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
numpy>=1.24.0
PyYAML>=6.0
pytest>=7.0.0
asyncpg>=0.27.0